+++ Source/ToolCore/Assets/ModelImporter.h
@@ -60,8 +60,8 @@ public:
 
     virtual void SetDefaults();
 
-    float GetScale() { return scale_; }
-    void SetScale(float scale) {scale_ = scale; }
+    double GetScale() { return scale_; }
+    void SetScale(double scale) {scale_ = scale; }
 
     bool GetImportAnimations() { return importAnimations_; }
     void SetImportAnimations(bool importAnimations) { importAnimations_ = importAnimations; }
@@ -88,7 +88,7 @@ protected:
     virtual bool SaveSettingsInternal(JSONValue& jsonRoot);
 
 
-    float scale_;
+    double scale_;
     bool importAnimations_;
     Vector<SharedPtr<AnimationImportInfo>> animationInfo_;
